requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
Deprecated==1.2.13
Django==2.2.2
django-cors-headers==3.1.0
django-filter==2.4.0
django-redis==5.0.0
django-simpleui==2022.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
idna==3.3
importlib-metadata==4.12.0
kombu==5.2.4
mysqlclient==2.1.1
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
pycparser==2.21
pycryptodomex==3.9.4
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==3.0.9
python-alipay-sdk==3.0.4
pytz==2022.1
qcloudsms-py==0.1.4
qiniu==7.9.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.1
